Refactor overlay generation into dedicated hook
